Begin work on unit testing description parsing
