+++ package.json
@@ -10,6 +10,7 @@
   "packageManager": "yarn@3.6.4",
   "devDependencies": {
     "@types/follow-redirects": "^1.14.2",
+    "@types/html-to-text": "^9.0.4",
     "@types/jest": "^29.5.5",
     "@types/node": "^20.8.4",
     "@typescript-eslint/eslint-plugin": "^6.7.4",
@@ -31,6 +32,7 @@
     "@types/js-yaml": "^4.0.9",
     "dotenv": "^16.3.1",
     "googleapis": "^129.0.0",
+    "html-to-text": "^9.0.5",
     "js-yaml": "^4.1.0"
   }
 }
