+++ src/classes/SlackChannel.ts
